+++ app/app.py
@@ -2,6 +2,7 @@ from flask import Flask, render_template, request, redirect, url_for
 
 import service as service
 import cmnUtils as cmnUtils
+import imgUtils as imgUtils
 
 app = Flask(__name__)
 
@@ -16,15 +17,14 @@ def root():
 def post():
     if request.method == 'POST':
         input_kana = cmnUtils.convert_new_line(request.form['input_kana'])
-        ## TODO:改行がエラー判定される
-        is_downloadable = service.can_downloadable(input_kana)
+        is_downloadable = imgUtils.can_downloadable(input_kana)
         converted_input_list = service.converted_kana_to_oshite(input_kana)
 
         # rendering for home.html.
         return render_template('home.html',
                                input_kana=input_kana,
                                converted_input_list=converted_input_list,
-                               fileType=service.FILE_TYPE_PNG,
+                               fileType=imgUtils.FILE_TYPE_PNG,
                                can_downloadable=is_downloadable
                                )
     else:  # error redirect.
@@ -41,8 +41,7 @@ def home():
 def download():
     if request.method == 'POST':
         input_kana = cmnUtils.convert_new_line(request.json['input_kana'])
-        download_file = service.download_image(input_kana)
-        return download_file
+        return service.download_image(input_kana)
     else:  # error redirect.
         return redirect(url_for('home'))
