+++ app/templates/home.html
@@ -58,7 +58,7 @@
     {% if can_downloadable %}
     <button id="oshite__img__download" type="button" class="btn home__download__btn" >画像ダウンロード</button>
     {% else %}
-    <input type="button" value="画像ダウンロード" disabled />
+    <button type="button" class="btn home__download__btn" disabled>画像ダウンロード</button>
     {% endif %}
     <p class="oshite__not__convert__notice">※ひらがな文字以外を入力された場合、変換できません。そのままの文字を灰色で出力します。</p>
     <p class="oshite__not__convert__notice">※ひらがな文字以外を入力された場合、画像ダウンロードできません。</p>
